refactor(loop/rendering): split rendering.py under the 500-LOC gate (#1058)
`src/teatree/loop/rendering.py` was 603 LOC on `origin/main`, already
over the `scripts/hooks/check_module_health.py` ceiling (max 500). The
gate only fires on threshold-crossing, so the file shipped over-budget;
any feature branch adding one line to it would then trip the gate. Same
systemic blocker class as #1054 (tick split) and #1056 (review split).

Behaviour-preserving split into two focused sibling modules by concern,
keeping the public import surface unchanged:

* `rendering_classification.py` — signal → typed ref, per-overlay
  bucketing, cross-scanner dedup. Contains `_classify_actions`,
  `_ClassifiedActions`, `_dedup_in_order`, `_dedup_classified`,
  `_issue_ref_from`, `_pr_ref`, `_str_field`, `_str_list_field`,
  `_is_url`, `_ticket_number_from_url`, the `Payload` type alias.
* `rendering_zones.py` — per-zone line builders. Contains
  `_populate_overlay_zones`, `_render_ticket_line`, `_render_pr_group`,
  `_render_action_line`, `_disposition_parts`, `_running_tasks_lines`,
  `_link`, `_is_pr_url`, the noise-state / PR-url constants.

`rendering.py` keeps the `zones_for` orchestrator and the
`_populate_availability_anchor` fail-open helper, and re-exports
(`__all__` + module-level imports) every name existing consumers and
tests import (`zones_for`, `_classify_actions`, `_issue_ref_from`,
`_IssueRef`, `_OverlayActionRefs`, `_PRRef`, `_render_action_line`,
`_render_pr_group`). Function bodies are unchanged — pure move.

Post-split LOC (per `check_module_health.py` counting, all < 500):

* `rendering.py`                  70
* `rendering_classification.py`  184
* `rendering_zones.py`           278

The rendering/statusline/tick test suite passes unchanged (652 of the
focused subset; 5696 in the full suite, 94.01% coverage).
