+++ src/teatree/loop/rendering.py
@@ -4,558 +4,41 @@ Split out of :mod:`teatree.loop.tick` so each module owns one concern: tick is
 the orchestrator (scan → dispatch → render), rendering is the formatter
 (classify actions per overlay, render anchor / action_needed / in_flight rows
 with OSC8 links, filter stale anchors).
-"""
 
-import re
-from collections.abc import Mapping
-from dataclasses import dataclass, field
-from typing import Any
+This module is the thin top-level orchestrator. The two concerns it
+coordinates live in focused sub-modules so each stays under the module
+health ceiling: ``rendering_classification`` turns each dispatched signal
+into a typed ref and buckets/dedups per overlay (``_classify_actions``),
+and ``rendering_zones`` turns the classified buckets into per-zone
+statusline rows (``_populate_overlay_zones`` and the line builders).
+
+The names re-exported below keep ``from teatree.loop.rendering import X``
+working for every existing consumer and test after the split.
+"""
 
 from teatree.loop.dispatch import DispatchAction
 from teatree.loop.pr_ticket_index import build_ticket_index
-from teatree.loop.rendering_dms import DmRef as _DmRef
-from teatree.loop.rendering_dms import dm_ref_from as _dm_ref_from
-from teatree.loop.rendering_dms import is_dm_action as _is_dm_action
-from teatree.loop.rendering_dms import render_dm_line as _render_dm_line
-from teatree.loop.rendering_items import (
-    _IssueRef,
-    _LinkCtx,
-    _OverlayActionRefs,
-    _PRRef,
-    _ReassignRef,
-    _render_canonical_item,
+from teatree.loop.rendering_classification import _ClassifiedActions, _classify_actions, _issue_ref_from
+from teatree.loop.rendering_items import _IssueRef, _OverlayActionRefs, _PRRef
+from teatree.loop.rendering_zones import (
+    _populate_overlay_zones,
+    _render_action_line,
+    _render_pr_group,
+    _running_tasks_lines,
 )
-from teatree.loop.statusline import StatuslineEntry, StatuslineZones, _hyperlink, colorize_enabled, plain_link
+from teatree.loop.statusline import StatuslineZones, colorize_enabled
 
-# DispatchAction payloads are `dict[str, Any]` by contract (see dispatch.py).
-# Renderer-side reads only ever look up scalar fields, so we narrow to
-# Mapping[str, Any] at the function signatures to keep the module health
-# gate happy without inventing a TypedDict per signal shape.
-type Payload = Mapping[str, Any]
-
-_DISPOSITION_LABELS: dict[str, str] = {
-    "issue_closed": "closed",
-    "unassigned": "reassigned",
-    "label_removed": "label-removed",
-}
-
-# Summary prefixes the reviewer-prs scanner stamps on its signals (see
-# ``scanners/reviewer_prs.py``): ``Review needed:`` for
-# ``reviewer_pr.new_sha``/``unreviewed`` and ``Approval dismissed:`` for
-# ``reviewer_pr.approval_dismissed``. ``_pr_ref`` keys the URL-tail iid
-# derivation off these so every reviewer dual-dispatch renders as a
-# clickable ``!N`` ref, while non-reviewer PR-URL signals keep their
-# human-readable generic line.
-_REVIEWER_SUMMARY_PREFIXES: tuple[str, ...] = ("Review needed:", "Approval dismissed:")
-
-
-def _is_url(text: object) -> bool:
-    return isinstance(text, str) and text.startswith(("http://", "https://"))
-
-
-def _link(text: str, url: object, *, colorize: bool) -> str:
-    if isinstance(url, str) and url.startswith(("http://", "https://")):
-        return _hyperlink(text, url) if colorize else plain_link(text, url)
-    return text
-
-
-def _pr_ref(action: DispatchAction) -> _PRRef | None:
-    payload = action.payload if isinstance(action.payload, dict) else {}
-    url = payload.get("url", "")
-    iid = payload.get("iid")
-    if not isinstance(iid, int) or iid == 0:
-        # Reviewer-pr signals don't ship `iid` in the payload but the MR URL
-        # ends with the numeric ref — derive it so the row renders as `!N`
-        # under the right overlay. Match every reviewer summary form (the
-        # scanner emits ``Review needed:`` for new_sha/unreviewed and
-        # ``Approval dismissed:`` for approval_dismissed) without hijacking
-        # other PR-URL-bearing signals (e.g. ``my_pr.open``), which keep
-        # their human-readable generic line.
-        if isinstance(url, str) and action.detail.startswith(_REVIEWER_SUMMARY_PREFIXES):
-            tail = _ticket_number_from_url(url)
-            if tail.isdigit():
-                return _PRRef(iid=int(tail), url=url, annotation="review")
-        return None
-    draft_count = payload.get("draft_count")
-    status = payload.get("status", "")
-    if isinstance(draft_count, int) and draft_count > 0:
-        return _PRRef(iid=iid, url=url, annotation=f"{draft_count} notes")
-    if status in {"failed", "failure", "error"}:
-        return _PRRef(iid=iid, url=url, annotation=f"pipeline {status}")
-    return _PRRef(iid=iid, url=url, annotation="")
-
-
-def _ticket_number_from_url(url: str) -> str:
-    match = re.search(r"(\d+)(?:/?)$", url)
-    return match.group(1) if match else ""
-
-
-def _render_pr_group(
-    overlay: str,
-    refs: list[_PRRef],
-    *,
-    ticket_index: dict[str, str] | None = None,
-    colorize: bool,
-) -> str:
-    """Render a flat list of PR refs, grouped per parent ticket when known."""
-    prefix = f"[{overlay}] " if overlay else ""
-    if not refs:
-        return ""
-    by_ticket: dict[str, list[_PRRef]] = {}
-    orphans: list[_PRRef] = []
-    for ref in refs:
-        parent = (ticket_index or {}).get(ref.url, "")
-        if parent and parent != str(ref.iid):
-            by_ticket.setdefault(parent, []).append(ref)
-        else:
-            orphans.append(ref)
-
-    def _label(ref: _PRRef) -> str:
-        text = f"!{ref.iid}"
-        if ref.annotation:
-            text += f" ({ref.annotation})"
-        return _link(text, ref.url, colorize=colorize)
-
-    chunks: list[str] = []
-    for tnum in sorted(by_ticket):
-        bucket = " ".join(_label(r) for r in by_ticket[tnum])
-        chunks.append(f"#{tnum}: {bucket}")
-    if orphans:
-        chunks.append(" · ".join(_label(r) for r in orphans))
-    return f"{prefix}{' · '.join(chunks)}"
-
-
-@dataclass(slots=True)
-class _ClassifiedActions:
-    disposition_refs: dict[str, dict[str, list[_IssueRef]]] = field(default_factory=dict)
-    reassign_refs: dict[str, list[_ReassignRef]] = field(default_factory=dict)
-    stale_refs: dict[str, list[_IssueRef]] = field(default_factory=dict)
-    ready_refs: dict[str, list[_IssueRef]] = field(default_factory=dict)
-    action_prs: dict[str, list[_PRRef]] = field(default_factory=dict)
-    inflight_prs: dict[str, list[_PRRef]] = field(default_factory=dict)
-    # ``(ticket_number, state, issue_url, title)`` — ``title`` is the cached
-    # tracker title (``ticket.extra["issue_title"]``); empty when the
-    # scanner has no title yet. Renderer uses it for the canonical
-    # ``#N (short desc) (!M)`` item shape (#1015).
-    active_tickets: dict[str, list[tuple[str, str, str, str]]] = field(default_factory=dict)
-    dms: dict[str, list[_DmRef]] = field(default_factory=dict)
-    other: list[tuple[str, StatuslineEntry]] = field(default_factory=list)
-
-
-_TITLE_FALLBACK_LEN = 32
-
-
-def _issue_ref_from(
-    *,
-    url: str = "",
-    issue_url: str = "",
-    ticket_number: str = "",
-    title: str = "",
-) -> _IssueRef:
-    # Dispositions emit ``issue_url``; my-pr and ready emit ``url``. Honour
-    # whichever the source provided so the renderer never collapses a real
-    # ticket into a bare ``?`` token. Falls back to ``ticket_number`` or a
-    # title slug when no usable URL is available.
-    url_str = url or issue_url
-    number = _ticket_number_from_url(url_str)
-    if number:
-        return _IssueRef(label=f"#{number}", url=url_str, title=title)
-    if ticket_number:
-        return _IssueRef(
-            label=f"#{ticket_number}",
-            url=url_str if _is_url(url_str) else "",
-            title=title,
-        )
-    if title:
-        snippet = title if len(title) <= _TITLE_FALLBACK_LEN else title[: _TITLE_FALLBACK_LEN - 3] + "…"
-        return _IssueRef(label=snippet, url=url_str, title=title)
-    return _IssueRef(label="?", url=url_str, title=title)
-
-
-def _str_field(payload: Payload, key: str) -> str:
-    value = payload.get(key)
-    return value if isinstance(value, str) else ""
-
-
-def _str_list_field(payload: Payload, key: str) -> tuple[str, ...]:
-    value = payload.get(key)
-    if not isinstance(value, list):
-        return ()
-    return tuple(item for item in value if isinstance(item, str))
-
-
-def _classify_actions(actions: list[DispatchAction]) -> _ClassifiedActions:
-    c = _ClassifiedActions()
-    for action in actions:
-        payload = action.payload if isinstance(action.payload, dict) else {}
-        url_str = _str_field(payload, "url")
-        overlay = _str_field(payload, "overlay")
-        prefix = f"[{overlay}] " if overlay else ""
-
-        if action.kind != "statusline":
-            continue
-        state = payload.get("state")
-        ticket_number = payload.get("ticket_number")
-        if action.zone == "anchors" and isinstance(state, str) and isinstance(ticket_number, str):
-            issue_url = _str_field(payload, "issue_url")
-            title = _str_field(payload, "title")
-            c.active_tickets.setdefault(overlay, []).append((ticket_number, state, issue_url, title))
-            continue
-        if payload.get("stale") is True:
-            c.stale_refs.setdefault(overlay, []).append(
-                _issue_ref_from(
-                    issue_url=_str_field(payload, "issue_url"),
-                    ticket_number=_str_field(payload, "ticket_number"),
-                ),
-            )
-            continue
-        reason = payload.get("reason")
-        if isinstance(reason, str):
-            ref = _issue_ref_from(
-                url=_str_field(payload, "url"),
-                issue_url=_str_field(payload, "issue_url"),
-                ticket_number=_str_field(payload, "ticket_number"),
-                title=_str_field(payload, "title"),
-            )
-            old_owner = _str_field(payload, "old_owner")
-            new_owners = _str_list_field(payload, "new_owners")
-            if reason == "unassigned" and old_owner and new_owners:
-                c.reassign_refs.setdefault(overlay, []).append(
-                    _ReassignRef(ref=ref, old_owner=old_owner, new_owners=new_owners),
-                )
-                continue
-            c.disposition_refs.setdefault(overlay, {}).setdefault(reason, []).append(ref)
-            continue
-        if action.zone == "action_needed" and action.detail.startswith("Ready to start:"):
-            c.ready_refs.setdefault(overlay, []).append(
-                _issue_ref_from(
-                    url=_str_field(payload, "url"),
-                    issue_url=_str_field(payload, "issue_url"),
-                    ticket_number=_str_field(payload, "ticket_number"),
-                    title=_str_field(payload, "title"),
-                ),
-            )
-            continue
-        if _is_dm_action(action, payload):
-            c.dms.setdefault(overlay, []).append(_dm_ref_from(payload))
-            continue
-        ref = _pr_ref(action)
-        if ref is not None:
-            bucket = c.action_prs if action.zone == "action_needed" else c.inflight_prs
-            bucket.setdefault(overlay, []).append(ref)
-            continue
-        c.other.append((action.zone, StatuslineEntry(text=f"{prefix}{action.detail}", url=url_str)))
-    _dedup_classified(c)
-    return c
-
-
-def _dedup_in_order[T](items: list[T]) -> list[T]:
-    """Return *items* with duplicates dropped, first-occurrence wins.
-
-    Statusline rows must be stable across ticks: two scanners surfacing
-    the same observation (same ``ticket.active`` row, same PR seen by both
-    ``MyPrsScanner`` and ``ReviewerPrsScanner``, same ``ticket.stale`` row
-    emitted again on the next sweep) must collapse to one ref. Order is
-    preserved so the user sees a deterministic line.
-    """
-    seen: set[T] = set()
-    out: list[T] = []
-    for item in items:
-        if item in seen:
-            continue
-        seen.add(item)
-        out.append(item)
-    return out
-
-
-def _dedup_classified(c: _ClassifiedActions) -> None:
-    """Collapse duplicate refs in every per-overlay bucket.
-
-    The dispatch layer is fan-out by signal (one signal per scanner that
-    saw it); the renderer is the boundary that must present each observed
-    thing exactly once. Doing the dedup here keeps every downstream
-    renderer simple — they can assume the input list has no duplicates.
-    """
-    for overlay, tickets in list(c.active_tickets.items()):
-        c.active_tickets[overlay] = _dedup_in_order(tickets)
-    for overlay, refs in list(c.stale_refs.items()):
-        c.stale_refs[overlay] = _dedup_in_order(refs)
-    for overlay, refs in list(c.ready_refs.items()):
-        c.ready_refs[overlay] = _dedup_in_order(refs)
-    for overlay, refs in list(c.action_prs.items()):
-        c.action_prs[overlay] = _dedup_in_order(refs)
-    for overlay, refs in list(c.inflight_prs.items()):
-        c.inflight_prs[overlay] = _dedup_in_order(refs)
-    for overlay, reass in list(c.reassign_refs.items()):
-        c.reassign_refs[overlay] = _dedup_in_order(reass)
-    for overlay, dms in list(c.dms.items()):
-        c.dms[overlay] = _dedup_in_order(dms)
-    for reason_map in c.disposition_refs.values():
-        for reason, refs in list(reason_map.items()):
-            reason_map[reason] = _dedup_in_order(refs)
-
-
-# States that should not surface as anchors: terminal/post-PR states plus
-# ``closed`` which isn't a valid FSM value but turns up in old data.
-_NOISE_STATES = frozenset(
-    {
-        "not_started",
-        "merged",
-        "delivered",
-        "shipped",
-        "in_review",
-        "retrospected",
-        "closed",
-    },
-)
-_MAX_PER_STATE = 5
-
-# Tickets whose ``issue_url`` matches one of these are treated as PR-backed.
-# A PR-backed ticket that doesn't appear in the live PR set (action_needed
-# or in_flight) is considered stale — its remote MR has likely been merged
-# or closed but the local FSM never advanced.
-_PR_URL_RE = re.compile(r"/(?:merge_requests|pull|pulls)/\d+/?$")
-
-
-def _is_pr_url(url: str) -> bool:
-    return bool(_is_url(url) and _PR_URL_RE.search(url))
-
-
-def _render_ticket_line(
-    overlay: str,
-    tickets: list[tuple[str, str, str, str]],
-    pr_map: dict[str, list[_PRRef]],
-    *,
-    live_pr_urls: set[str] | None = None,
-    colorize: bool,
-) -> str:
-    """Render the per-overlay anchor line — one row per state group.
-
-    Every state line (``ready:``, ``started:``, ``tested:``, …) uses the
-    same canonical item shape: ``#N (short desc) (!M1, !M2)`` where the
-    description is the cached tracker title truncated to
-    ``_ITEM_DESC_LEN`` and the MR refs are comma-separated and clickable
-    (#1015). The PR group falls back to a space-separated form when no
-    description is present, but in the canonical path every number is a
-    hyperlink.
-
-    ``pr_map`` is the overlay's MR-iid → child-refs map; entries appear
-    here either because the MR's iid equals the ticket number (legacy
-    shape), or because the caller pre-bucketed by parent ticket number
-    via ``ticket_index`` (canonical shape).
-    """
-    prefix = f"[{overlay}] " if overlay else ""
-    live = live_pr_urls or set()
-    by_state: dict[str, list[str]] = {}
-    for num, state, url, title in tickets:
-        if state in _NOISE_STATES:
-            continue
-        if _is_pr_url(url) and url not in live:
-            continue
-        by_state.setdefault(state, []).append(
-            _render_canonical_item(
-                label=f"#{num}",
-                url=url,
-                title=title,
-                child_refs=pr_map.get(num, []),
-                ctx=_LinkCtx(colorize=colorize, link=_link),
-            ),
-        )
-    if not by_state:
-        return ""
-    groups: list[str] = []
-    for state, items in by_state.items():
-        shown = items[:_MAX_PER_STATE]
-        overflow = len(items) - len(shown)
-        label = " ".join(shown)
-        if overflow > 0:
-            label += f" (+{overflow})"
-        groups.append(f"{state}: {label}")
-    return f"{prefix}{' · '.join(groups)}"
-
-
-def _disposition_parts(action_refs: _OverlayActionRefs, *, colorize: bool) -> list[str]:
-    """Render the issue-disposition rows for one overlay.
-
-    Covers generic dispositions, the explicit ``reassigned (from → to)``
-    transition, and the collapsed ``N stale`` row. Each is one concise part
-    with linked refs — the stale row in particular folds every stale ticket
-    for the overlay into a single line.
-    """
-    parts: list[str] = []
-    for reason, refs in action_refs.disposition_refs.items():
-        label = _DISPOSITION_LABELS.get(reason, reason)
-        items = " ".join(_link(r.label, r.url, colorize=colorize) for r in refs)
-        parts.append(f"{label}: {items}")
-    for rr in action_refs.reassign_refs:
-        to = ", ".join(rr.new_owners)
-        ref_link = _link(rr.ref.label, rr.ref.url, colorize=colorize)
-        parts.append(f"reassigned (from {rr.old_owner} → to {to}): {ref_link}")
-    if action_refs.stale_refs:
-        stale = action_refs.stale_refs
-        items = " ".join(_link(r.label, r.url, colorize=colorize) for r in stale)
-        parts.append(f"{len(stale)} stale: {items}")
-    return parts
-
-
-def _render_action_line(
-    overlay: str,
-    action_refs: _OverlayActionRefs,
-    *,
-    ticket_index: dict[str, str] | None = None,
-    colorize: bool,
-) -> str:
-    prefix = f"[{overlay}] " if overlay else ""
-    prs_by_ticket: dict[str, list[_PRRef]] = {}
-    for ref in action_refs.pr_refs:
-        parent = (ticket_index or {}).get(ref.url, "")
-        if parent and parent != str(ref.iid):
-            prs_by_ticket.setdefault(parent, []).append(ref)
-    consumed_pr_urls: set[str] = set()
-
-    parts: list[str] = _disposition_parts(action_refs, colorize=colorize)
-    if action_refs.ready_refs:
-        items: list[str] = []
-        for ref in action_refs.ready_refs:
-            number = ref.label.lstrip("#")
-            prs = prs_by_ticket.get(number, [])
-            items.append(
-                _render_canonical_item(
-                    label=ref.label,
-                    url=ref.url,
-                    title=ref.title,
-                    child_refs=prs,
-                    ctx=_LinkCtx(colorize=colorize, link=_link),
-                ),
-            )
-            consumed_pr_urls.update(p.url for p in prs)
-        parts.append(f"ready: {' '.join(items)}")
-    if action_refs.pr_refs:
-        remaining = [r for r in action_refs.pr_refs if r.url not in consumed_pr_urls]
-        if remaining:
-            parts.insert(
-                0,
-                _render_pr_group(overlay, remaining, ticket_index=ticket_index, colorize=colorize).removeprefix(prefix),
-            )
-    if not parts:
-        return ""
-    return f"{prefix}{' · '.join(parts)}"
-
-
-def _running_tasks_lines() -> list[str]:
-    """Render the ``[ov] agents: <phase> · <phase>`` row from CLAIMED tasks.
-
-    Skips tasks with a blank ``phase`` (the column is ``blank=True`` and
-    the default is ``""``) so the renderer never produces a phantom
-    ``agents:  · coding`` with a double space the operator can't act on.
-    Dedupes phase names per overlay: two parallel implementers both at
-    ``coding`` collapse to one ``agents: coding`` ref — concurrency is
-    implied by the loop, not by visually repeating the same word.
-    Suppresses the whole row when no claimed task contributes a phase.
-    """
-    from django.apps import apps  # noqa: PLC0415
-
-    try:
-        task_model = apps.get_model("core", "Task")
-        claimed = (
-            task_model.objects.filter(status="claimed")
-            .select_related("ticket")
-            .only("phase", "ticket__overlay", "ticket__issue_url")
-        )
-        by_overlay: dict[str, list[str]] = {}
-        for task in claimed:
-            phase = (task.phase or "").strip()
-            if not phase:
-                continue
-            overlay = task.ticket.overlay or ""
-            by_overlay.setdefault(overlay, []).append(phase)
-    except Exception:  # noqa: BLE001
-        return []
-    lines: list[str] = []
-    for overlay, phases in sorted(by_overlay.items()):
-        unique_phases = _dedup_in_order(phases)
-        if not unique_phases:
-            continue
-        prefix = f"[{overlay}] " if overlay else ""
-        lines.append(f"{prefix}agents: {' · '.join(unique_phases)}")
-    return lines
-
-
-def _populate_overlay_zones(
-    zones: StatuslineZones,
-    c: _ClassifiedActions,
-    *,
-    ticket_index: dict[str, str],
-    colorize: bool,
-) -> None:
-    all_overlays = sorted(
-        {
-            *c.active_tickets,
-            *c.action_prs,
-            *c.disposition_refs,
-            *c.reassign_refs,
-            *c.stale_refs,
-            *c.ready_refs,
-            *c.inflight_prs,
-            *c.dms,
-        },
-    )
-
-    # ``pr_map`` keys are ticket numbers: each MR is bucketed under its
-    # parent ticket (resolved via ``ticket_index`` — ``Closes #N`` footer
-    # parsing in ``pr_ticket_index``). Legacy fall-through: when no parent
-    # is known, the MR is still keyed by its own iid so a ticket-number ==
-    # iid coincidence keeps rendering.
-    all_pr_refs: dict[str, dict[str, list[_PRRef]]] = {}
-    for overlay_key in all_overlays:
-        for refs in (c.action_prs.get(overlay_key, []), c.inflight_prs.get(overlay_key, [])):
-            for ref in refs:
-                parent = ticket_index.get(ref.url, "")
-                key = parent if parent and parent != str(ref.iid) else str(ref.iid)
-                all_pr_refs.setdefault(overlay_key, {}).setdefault(key, []).append(ref)
-
-    live_pr_urls_by_overlay: dict[str, set[str]] = {}
-    for overlay_key in all_overlays:
-        live = {r.url for r in c.action_prs.get(overlay_key, []) if r.url}
-        live |= {r.url for r in c.inflight_prs.get(overlay_key, []) if r.url}
-        live_pr_urls_by_overlay[overlay_key] = live
-
-    for overlay_key in all_overlays:
-        pr_map = all_pr_refs.get(overlay_key, {})
-        ticket_line = _render_ticket_line(
-            overlay_key,
-            c.active_tickets.get(overlay_key, []),
-            pr_map,
-            live_pr_urls=live_pr_urls_by_overlay.get(overlay_key, set()),
-            colorize=colorize,
-        )
-        if ticket_line:
-            zones.anchors.append(ticket_line)
-
-        dm_line = _render_dm_line(overlay_key, c.dms.get(overlay_key, []), link=_link, colorize=colorize)
-        if dm_line:
-            zones.anchors.append(dm_line)
-
-        action_line = _render_action_line(
-            overlay_key,
-            _OverlayActionRefs(
-                pr_refs=c.action_prs.get(overlay_key, []),
-                disposition_refs=c.disposition_refs.get(overlay_key, {}),
-                ready_refs=c.ready_refs.get(overlay_key, []),
-                reassign_refs=c.reassign_refs.get(overlay_key, []),
-                stale_refs=c.stale_refs.get(overlay_key, []),
-            ),
-            ticket_index=ticket_index,
-            colorize=colorize,
-        )
-        if action_line:
-            zones.action_needed.append(action_line)
-
-        inflight_refs = c.inflight_prs.get(overlay_key, [])
-        if inflight_refs:
-            zones.in_flight.append(
-                _render_pr_group(overlay_key, inflight_refs, ticket_index=ticket_index, colorize=colorize)
-            )
+__all__ = [
+    "_ClassifiedActions",
+    "_IssueRef",
+    "_OverlayActionRefs",
+    "_PRRef",
+    "_classify_actions",
+    "_issue_ref_from",
+    "_render_action_line",
+    "_render_pr_group",
+    "zones_for",
+]
 
 
 def zones_for(actions: list[DispatchAction], *, colorize: bool | None = None) -> StatuslineZones:
